refactor(server): extract Codex app-server transport
